refactor: isolate yara rule path resolution
